Fully implemented cullable lights and optimized lighting extensively.
Used bounding boxes for light culling. Added a feature for the frustum to be cones instead of regular camera frustums in order to improve optimization. Minimized HLSL code calculations and moved taxing calculations to the C++ side for better performance. As a result,  we achieved a 250% performance boost during testing.